+++ RxGRDB/ValueObservation+Rx.swift
@@ -28,13 +28,14 @@ extension Reactive where Base: _ValueObservationProtocol {
     ///             print("Fresh players: \(players)")
     ///         })
     ///
-    /// By default, all values are emitted on the main dispatch queue. If you
-    /// give a *scheduler*, values are emitted on that scheduler.
+    /// All elements are emitted on the main queue by default, unless you
+    /// provide a specific `scheduler`.
     ///
-    /// If you set *startImmediately* to true (the default value), the first
-    /// element is emitted right upon subscription. It is *synchronously*
-    /// emitted if and only if the observable is subscribed on the main queue,
-    /// and is given a nil *scheduler* argument:
+    /// If you set `startImmediately` to true (the default value), the first
+    /// element is emitted immediately, from the current database state.
+    /// Furthermore, this first element is emitted *synchronously* if and only
+    /// if the observable is subscribed on the main queue, and is given a nil
+    /// `scheduler` argument:
     ///
     ///     // on the main queue
     ///     observation.rx